requirements_dev.txt
setuptools
pre-commit
ml_collections
absl-py
pytest
tox
black
pytest
pylint
mkdocs
mkdocs-material
mkdocstrings[python]
mknotebooks
bump2version
wandb